% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_sex_specific_exclusions.R
\name{apply_sex_specific_exclusions}
\alias{apply_sex_specific_exclusions}
\title{Apply sex specific exclusions}
\usage{
apply_sex_specific_exclusions(phewas_df, name_of_sex_column)
}
\arguments{
\item{phewas_df}{Input data frame}

\item{name_of_sex_column}{Name of the sex column in the input data frame}
}
\value{
phewas data frame with sex exclusions applied
}
\description{
This function applies sex-specific exclusions for male and female reproductive system procedures
FRS___ = female reproductive system
MRS___ = male reproductive system
PGN___ = pregnancy related procedures
}
