% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ccsr_phenotypes.R
\name{create_ccsr_phenotypes}
\alias{create_ccsr_phenotypes}
\title{Map source codes to CCSR categories}
\usage{
create_ccsr_phenotypes(
  source_codes_df,
  mapping_file = ProcWAS::icd10_cpt4_source_to_ccsr
)
}
\arguments{
\item{source_codes_df}{Input data frame with source codes}

\item{mapping_file}{Mapping file to be used to map source icd10, cpt codes to ccsr categories. Default is the icd10_cpt4_source_to_ccsr mapping file from the package.}
}
\value{
Data frame with CCSR categories mapped and reshaped wider
}
\description{
This function maps the source procedure codes to CCSR categories and reshapes the data frame wider in preparation for the ProcWAS. The input data frame must have 4 columns with "id" representing the patient id, "vocabulary" representing one of "cpt_hcpcs" or "icd10", "code" representing the procedure code, and "date" representing the date the procedure was performed.
}
