% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pheno_with_cov.R
\name{merge_pheno_with_cov}
\alias{merge_pheno_with_cov}
\title{Merge phenotypes with covariates}
\usage{
merge_pheno_with_cov(
  phenotypes_df,
  covariates_df,
  append_cov_without_pheno_as_controls = TRUE
)
}
\arguments{
\item{phenotypes_df}{Input phenotypes data frame}

\item{covariates_df}{Input covariates data frame}

\item{append_cov_without_pheno_as_controls}{Should individuals in the covariate data but not in the phenotypes data be appended as controls to the phenotypes data Default is TRUE}
}
\value{
Merged phenotypes and covariates data frame
}
\description{
This function merges the phenotype information (output of create_ccsr_phenotypes) with covariate information.
}
