% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pheno_with_full_pop_cov.R
\name{merge_pheno_with_full_pop_cov}
\alias{merge_pheno_with_full_pop_cov}
\title{Merge phenotypes with covariates of the full population}
\usage{
merge_pheno_with_full_pop_cov(
  phenotypes_df,
  covariates_df,
  append_cov_without_pheno_as_controls = TRUE
)
}
\arguments{
\item{phenotypes_df}{Input phenotypes data frame. This is the output of create_ccsr_phenotypes where each row represents a unique participant id ("id") and each column represents a procedure}

\item{covariates_df}{Input covariates data frame. This is a data frame with at least one column representing the participant id ("id") and additional columns to be included as covariates in the regression.}

\item{append_cov_without_pheno_as_controls}{Should individuals in the covariate data but not in the phenotypes data be appended as controls to the phenotypes data. Default is TRUE}
}
\value{
Merged phenotypes and covariates data frame
}
\description{
This function joins the phenotype information with covariate information.
}
