% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_as_ext.R
\name{phe_as_ext}
\alias{phe_as_ext}
\title{ProcWAS using generalized linear model}
\source{
https://github.com/monikagrabowska/PedsPheWAS/blob/main/R/phe_as_ext.R

https://github.com/PheWAS/PheWAS/blob/master/R/phe_as_ext.R
}
\usage{
phe_as_ext(
  phe.gen,
  additive.genotypes = T,
  min.records = 20,
  return.models = F,
  confint.level = NA,
  factor.contrasts = NA,
  my.data,
  ...
)
}
\value{
phenotype-significance Manhattan plot to its parent function
}
\description{
Called within the phewas_ext() function to perform ProcWAS analysis using a generalized linear model (“glm”). For more details on arguments/outputs of PheWAS, see the phewas_ext() function.
}
