% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_ext.R
\name{phewas_ext}
\alias{phewas_ext}
\title{Performs ProcWAS
This function perofrms ProcWAS using generalized linear model with optional adjustment for user-specified covariates}
\source{
https://github.com/monikagrabowska/PedsPheWAS/blob/main/R/phewas_ext.R

https://github.com/PheWAS/PheWAS/blob/master/R/phewas_ext.R
}
\usage{
phewas_ext(
  phenotypes,
  genotypes,
  data,
  covariates = NA,
  cores = 1,
  additive.genotypes = T,
  return.models = F,
  min.records = 20,
  MASS.confint.level = NA
)
}
\arguments{
\item{phenotypes}{The names of the outcome variables (i.e., the procedures)}

\item{genotypes}{The names of the prediction variables (i.e., the genotype)}

\item{data}{A data frame containing all of the variables to be used in the ProcWAS analysis}

\item{covariates}{The names of the covariates}

\item{cores}{The number of cores to be used in the parallel socket cluster implementation.}

\item{additive.genotypes}{Identifies if additive genotypes are being supplied (default is TRUE).}

\item{return.models}{Returns a list of the complete models (default is FALSE).}

\item{min.records}{The minimum number of records necessary to perform a test. For logistic regression, there must be at least this number of cases and at least this number of controls. The default value is 20.}

\item{MASS.confint.level}{Calculates a confidence interval at the specified level. Logistic models will report OR CIs. Confidence intervals are not calculated by default.}
}
\value{
Returns a data frame containing the results of the PheWAS analysis (e.g., a data frame with the following columns: phenotype, snp, covariates, beta, SE, OR, p, type, n_total, n_cases, n_controls, HWE_p, allele_freq, n_no_snp, formula, expanded_formula, note).
}
\description{
Performs ProcWAS
This function perofrms ProcWAS using generalized linear model with optional adjustment for user-specified covariates
}
