% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotManhattan.R
\name{plotManhattan}
\alias{plotManhattan}
\title{ProcWAS Manhattan Plotting
Generates a PheWAS Manhattan plot with phenotypes on the x-axis and p-values on a -log10 scale on the y-axis. This function visualizes the results generated by the phewas_ext() function that performs the ProcWAS analysis.}
\source{
https://github.com/monikagrabowska/PedsPheWAS/blob/main/R/plotManhattan.R

https://github.com/PheWAS/PheWAS/blob/master/R/phewasManhattan.R
}
\usage{
plotManhattan(
  d,
  suggestive.line = 0.05,
  significant.line,
  annotate.level,
  y.axis.interval = 5,
  ...
)
}
\arguments{
\item{d}{This will typically be the output data frame from the phewas_ext() function (i.e., the PheWAS results), although this can also be a user-specified data frame with phenotype, OR, and p columns. The phenotype column should contain the phecodes, not the phecode descriptions. The name of the OR column needs to be capitalized (i.e., “OR” not “or”).}

\item{suggestive.line}{The p-value at which to draw a blue line to show a significance threshold. The default value is 0.05. An NA value results in no line.}

\item{significant.line}{The p-value at which to draw a red line to show an adjusted significance threshold. The default applies a Bonferroni correction. An NA value results in no line.}

\item{annotate.level}{The p-value threshold for annotating phenotypes on the Manhattan plot. The default value is that provided in the significant.line argument.}

\item{y.axis.interval}{The interval used for the y-axis of the Manhattan plot (default value = 5).}
}
\value{
This function returns a Manhattan Plot with a log(10) y axis and phenotypes as the x axis.
}
\description{
ProcWAS Manhattan Plotting
Generates a PheWAS Manhattan plot with phenotypes on the x-axis and p-values on a -log10 scale on the y-axis. This function visualizes the results generated by the phewas_ext() function that performs the ProcWAS analysis.
}
