% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcWAS-package.R
\docType{package}
\name{ProcWAS-package}
\alias{ProcWAS}
\alias{ProcWAS-package}
\title{ProcWAS: Procedure-wide Association Study}
\description{
This package adapts functions from the PheWAS package to perform a procedure-wide association study (ProcWAS).
}
\keyword{internal}
